+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v48.x / v49 (In progress)
+- PC sprites: Added Erdick and Link.
+
 ## v48 (2021-04-28)
 - Ancient Cave: Spawn plates around doors adjusted and hardlock glitch fixed.
 - Credits: EBSS AC 2021 winner DoctorSwellman added.